Add files via upload
The roadmap has 3 kinds of roads:
public roads, private roads that have a fee to be used, and construction roads that are under repair,
so the traffic through them is slow.
Program will receive as input a file with a description of the roadmap, the starting point
s, the destination e, the maximum number p of private roads that the path from s to e can use,
and the maximum number c of construction roads that the path from s to e can use. Your program
then will try to find a path from s to e as specified.
Every edge of the graph represents a road
and every node represents either the intersection of two roads or the end of a road. There are two
special nodes in this graph denoting the starting point s and the destination e